+++ CHANGELOG.md
@@ -18,6 +18,7 @@ v5.5.0
 
 **IncNavierStokesSolver**
 - Save BndElmtExpansion and avoid re-building (!1648)
+- Add Simo-advection and a switch for Simo-/Dong-advection to VCSImplicit (!1630)
 
 **FieldConvert**
 - Fix typo in user-guide and fix but for parallel-in-time FieldConvert (!1645)